Hormone signature now inherits \n
Before these changes, the signature was calculated line by line but without the line delimiter. Since the hormone class uses the line delimiter to divided the different fields, this was not acceptable. The payload "1234\n1\12" had the same signature like "1234\n11\n2" but not the same meaning. The definition still ignores the line delimiter. But the format of the definitions is JSON, which simply ignores line breaks. The transmitted information does not change by moving characters from one line to another.
Of course, this breaks all unit tests using precalculated payloads. I'm gonna fix this late ...