Definition resend interval is now adjustable
The definition is resend in an interval. The previous default was 1h. This was quite heavy since the hormone definition is emitted with retained flag and should be retained forever at the broker. Since everything in the Internet should have a timeout, the defintion might be removed after some time with no refresh.